requirements.txt
openai>=1.12.0
selenium>=4.16.0
webdriver-manager>=4.0.1
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
bibtexparser>=1.4.1
